move SparkHiveHadoopWriter to org.apache.spark package
